detection/requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
brotlipy==0.7.0
cachetools==5.3.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cmake==3.27.2
contourpy==1.0.7
cycler==0.11.0
datasets==2.12.0
dill==0.3.6
filelock==3.12.0
flask==2.3.3
flask-sqlalchemy==3.0.5
flatbuffers==23.3.3
fonttools==4.39.4
frozenlist==1.3.3
fsspec==2023.5.0
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.57.0
h11==0.14.0
huggingface-hub==0.14.1
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
itsdangerous==2.1.2
jax==0.4.8
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.1
kaggle==1.5.13
keras==2.12.0
kiwisolver==1.4.4
libclang==16.0.0
lit==16.0.6
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
ml-dtypes==0.1.0
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
networkx==3.1
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.1
pdf2image==1.16.3
Pillow==9.5.0
protobuf==4.23.0
psutil==5.9.5
pyarrow==12.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyMuPDF==1.22.3
PyMySQL==1.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2023.3
PyYAML==6.0
regex==2023.5.5
requests==2.31.0
requests-oauthlib==1.3.1
responses==0.18.0
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
sentencepiece==0.1.99
simple-websocket==0.10.1
six==1.16.0
smmap==5.0.0
sock==0.6.0
SQLAlchemy==2.0.20
sympy==1.12
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.12.0
termcolor==2.3.0
text-unidecode==1.3
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tokenizers==0.13.3
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.66.1
transformers==4.29.1
triton==2.0.0
typing-extensions==4.5.0
tzdata==2023.3
ultralytics==8.0.120
urllib3==2.0.4
websocket-client==0.57.0
werkzeug==2.3.7
wrapt==1.14.1
wsproto==1.2.0
xxhash==3.2.0
yarl==1.9.2
zipp==3.16.2
